Change CancelRecurringPaymentsProfileAction namespace
